generated_code/fixtures/tests/test_extractimportsfromfile_fixture.py
import pytest
from generated_code.fixtures.fixtures.extractimportsfromfile_fixture import fixture_mock_extractimportsfromfile
from code_autoeval.llm_model.imports.extract_imports_from_file import ExtractImportsFromFile
from pathlib import Path
from typing import Any, Dict
from generated_code.fixtures.fixtures.loggingfuncs_fixture import fixture_mock_loggingfuncs
from generated_code.fixtures.fixtures.commonattributes_fixture import fixture_mock_commonattributes
from generated_code.fixtures.fixtures.llmmodelattributes_fixture import fixture_mock_llmmodelattributes
from generated_code.fixtures.fixtures.initkwargs_fixture import fixture_mock_initkwargs
def test_mock_extractimportsfromfile(fixture_mock_extractimportsfromfile):
    assert isinstance(fixture_mock_extractimportsfromfile, ExtractImportsFromFile)
    assert hasattr(fixture_mock_extractimportsfromfile, '_read_in_original_code')
    assert callable(fixture_mock_extractimportsfromfile._read_in_original_code)
    assert hasattr(fixture_mock_extractimportsfromfile, 'extract_imports')
    assert callable(fixture_mock_extractimportsfromfile.extract_imports)
    assert hasattr(fixture_mock_extractimportsfromfile, 'find_original_code_and_imports')
    assert isinstance(fixture_mock_extractimportsfromfile.find_original_code_and_imports, classmethod)
    assert callable(fixture_mock_extractimportsfromfile.find_original_code_and_imports.__func__)
    assert hasattr(fixture_mock_extractimportsfromfile, 'coverage_result')
    assert hasattr(fixture_mock_extractimportsfromfile, 'file_path')
    assert isinstance(fixture_mock_extractimportsfromfile.file_path, Path)
    assert hasattr(fixture_mock_extractimportsfromfile, 'test_file_path')
    assert isinstance(fixture_mock_extractimportsfromfile.test_file_path, Path)
    assert hasattr(fixture_mock_extractimportsfromfile, 'absolute_path_from_root')
    assert isinstance(fixture_mock_extractimportsfromfile.absolute_path_from_root, Path)
    assert hasattr(fixture_mock_extractimportsfromfile, 'unique_imports_dict')
    assert isinstance(fixture_mock_extractimportsfromfile.unique_imports_dict, Dict)
    assert hasattr(fixture_mock_extractimportsfromfile, 'find_original_code_and_imports')
    assert isinstance(fixture_mock_extractimportsfromfile.find_original_code_and_imports, classmethod)